Add vault.loadManifest from context provider.
